level1

+++ Java/CodingTest/src/programmers/MathSqrt.java
@@ -7,7 +7,7 @@ public class MathSqrt {
     public static long solution(long n) {
         long answer = 0;
         if ((int)Math.sqrt(n) == Math.sqrt(n)) {
-            int temp = (int)Math.sqrt(n);
+            long temp = (long)Math.sqrt(n);
             temp++;
             answer = temp * temp;
         } else {
